второй получше вариант реализации
