+++ js/data.js
@@ -1,6 +1,3 @@
-
-import {getRandomInteger, getRandomArrayElement, getRandomId, getRandomText} from './util.js'
-
 const RANDOM_TEXT =
   'Всё отлично! В целом всё неплохо, но не всё. Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально. Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше. Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше. Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!';
 
@@ -68,30 +65,15 @@ const MAX_COMMENTS = 30;
 const MIN_AVATAR = 1;
 const MAX_AVATAR = 6;
 
-const getUserId = getRandomId();
-const getUserPhoto = getRandomId();
-const getCommentId = getRandomId();
-
-const createComment = () => ({
-  id: getCommentId(),
-  avatar: `img/avatar-${getRandomInteger(MIN_AVATAR, MAX_AVATAR)}.svg`,
-  message: getRandomText(RANDOM_TEXT),
-  name: getRandomArrayElement(NAMES),
-});
-
-const addComment = () => Array.from(
-  { length: getRandomInteger(MIN_COMMENTS, MAX_COMMENTS) },
-  createComment
-);
-
-const createUser = () => ({
-  id: getUserId(),
-  url: `photos/${getUserPhoto()}.jpg`,
-  description: getRandomArrayElement(PHOTO_DESCRIPTIONS),
-  likes: getRandomInteger(MIN_LIKES, MAX_LIKES),
-  comments: addComment(),
-});
-
-const addUser = () => Array.from({ length: MAX_USERS }, createUser);
-
-export {addUser}
+export {
+  RANDOM_TEXT,
+  NAMES,
+  PHOTO_DESCRIPTIONS,
+  MAX_USERS,
+  MAX_LIKES,
+  MIN_LIKES,
+  MIN_COMMENTS,
+  MAX_COMMENTS,
+  MIN_AVATAR,
+  MAX_AVATAR,
+};
